Refactor: 인증 로직을 `global/auth`로 이동
