+++ src/main/java/com/team7/retriever/global/auth/jwt/repository/TokenRepository.java
@@ -1,10 +1,10 @@
-package com.team7.retriever.auth.jwt.repository;
+package com.team7.retriever.global.auth.jwt.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.team7.retriever.auth.redis.Token;
+import com.team7.retriever.global.auth.redis.Token;
 
 public interface TokenRepository extends CrudRepository<Token, String> {
