+++ src/main/java/com/team7/retriever/global/auth/security/AdminAuthentication.java
@@ -1,16 +1,14 @@
-package com.team7.retriever.auth.security;
+package com.team7.retriever.global.auth.security;
 
 import java.util.Collection;
 
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.GrantedAuthority;
 
-public class MemberAuthentication extends UsernamePasswordAuthenticationToken {
+public class AdminAuthentication extends UsernamePasswordAuthenticationToken {
 
-	// 사용자 인증 객체 생성
-	public MemberAuthentication(Object principal, Object credentials,
+	public AdminAuthentication(Object principal, Object credentials,
 		Collection<? extends GrantedAuthority> authorities) {
 		super(principal, credentials, authorities);
 	}
 }
-
