+++ src/main/java/com/team7/retriever/global/auth/security/CustomJwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.team7.retriever.auth.security;
+package com.team7.retriever.global.auth.security;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
